Heavily refactor the Addr class.
This removes the ability to implicitly convert from an Addr to an
integer type (you must explicitly convert now), and removes the
ambiguous overloaded operators, as well as some operators that I think
we can live without until there's a pressing reason to add them back.

+++ src/Symbol.cpp
@@ -54,7 +54,7 @@ gtirb::Symbol::StorageKind Symbol::getStorageKind() const {
 
 void Symbol::toProtobuf(MessageType* Message) const {
   nodeUUIDToBytes(this, *Message->mutable_uuid());
-  Message->set_address(this->Address);
+  Message->set_address(static_cast<uint64_t>(this->Address));
   Message->set_name(this->Name);
   Message->set_storage_kind(static_cast<proto::StorageKind>(this->Storage));
   uuidToBytes(this->CodeReferent.getUUID(),
